fix: luxonis dataset converter missing attr

+++ tests/unittests/test_converters.py
@@ -271,7 +271,7 @@ class TestLuxonisDatasetConverter(TestYOLOConverter):
 
     def tearDown(self):
         shutil.rmtree(self.test_dir)
-        if LuxonisDataset.exists(self.dataset_name):
+        if hasattr(self, self.dataset_name) and LuxonisDataset.exists(self.dataset_name):
             dataset = LuxonisDataset(self.dataset_name)
             dataset.delete_dataset()
